pyproject.toml
[tool.poetry]
name = "retail-company-assistant"
version = "0.1.0"
description = "CLI-based retail analytics assistant prototype"
authors = ["CTW <ctw01670@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
langgraph = ">=0.2.0"
langchain-core = ">=0.3.0"
langchain-google-genai = ">=1.0.0"
langchain-ollama = ">=0.2.0"
google-cloud-bigquery = ">=3.13.0"
pandas = ">=2.0.0"
python-dotenv = ">=1.0.0"
db-dtypes = "1.2.0"
PyYAML = ">=6.0.1"
pytest = ">=8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
